Condensing session/suite/global entry
